PROJECTS/rag-chatbot-llama3.2-gemma7B/requirements.txt
streamlit
langchain==0.3.7
langchain-core==0.3.15
sentence-transformers
pypdf
python-docx
python-dotenv
huggingface-hub==0.13.4
langchain-community==0.3.5
faiss-cpu
textract
groq
langchain-groq
PyPDF2
langchain_google_genai
pydantic==2.7.4
